WIP: Test Case - Module TC 05 and more
