+++ tests/CommonFile/PreAction.js
@@ -4,7 +4,7 @@ const { changeNetwork } = require("./Action");
 
 async function PreProposal(page){
 
-    test.setTimeout(10 * 60 * 1000); //Extending Test Case timeout to 10 minutes
+    test.setTimeout(5 * 60 * 1000); //Extending Test Case timeout to 5 minutes
     
     await changeNetwork(page);   //PreConditions Open URL and Selecting the Ghost-net
