+++ tests/TC01/TC01-02_TokenValues.spec.js
@@ -5,7 +5,7 @@ const { TezosTestData } = require("../CommonFile/TestData");
 
 test("Test Case 2: Token Creation with Various Input Values", async ({ page, }) => {
 
-  test.setTimeout(7 * 60 * 1000); //Extending Test Case timeout to 7 minutes
+  test.setTimeout(5 * 60 * 1000); //Extending Test Case timeout to 5 minutes
 
   await changeNetwork(page); //PreConditions Open URL and selecting the Ghost-net
