+++ tests/TC02/TC02-02_AcceptDelegation.spec.js
@@ -5,7 +5,7 @@ const { TezosTestData } = require("../CommonFile/TestData");
 
 test('Test case 2: Accept Delegations', async ({ page }) => {
 
-    test.setTimeout(7 * 60 * 1000); //Extending Test Case timeout to 7 minutes
+    test.setTimeout(5 * 60 * 1000); //Extending Test Case timeout to 5 minutes
 
     await PreProposal(page);   //PreConditions Open URL and Open Mask DAO For Proposal
