+++ tests/TC03/TC03-03_CorrectVoteCountOnOffChainPoll.spec.js
@@ -4,7 +4,7 @@ const { pageLocators } = require("../CommonFile/Locator");
 
 test('Test case 3: Correct vote Count on Off Chain Poll ', async ({ page }) => {
 
-  test.setTimeout(7 * 60 * 1000); //Extending Test Case timeout to 7 minutes
+  test.setTimeout(5 * 60 * 1000); //Extending Test Case timeout to 5 minutes
 
   await PreProposal(page);   //PreConditions Open URL and Open Mask DAO For Proposal
