+++ tests/TC04/TC04-06_RemoveLambda.spec.js
@@ -9,7 +9,7 @@ test('Test case 6: Remove Lambda ',  async ({ page }) => {
 
     try {
 
-        test.setTimeout(7 * 60 * 1000); //Extending Test Case timeout to 7 minutes
+        test.setTimeout(5 * 60 * 1000); //Extending Test Case timeout to 5 minutes
 
         await page.click(pageLocators.OffChainPoll.cycle);  // Click on Running Cycle Status
