+++ tests/TC06/TC06-01_SuccessfulTokenUnStaking.spec.js
@@ -4,7 +4,7 @@ const { pageLocators } = require("../CommonFile/Locator");
 
 test('Test case 1: Successful Token UnStaking ', async ({ page }) => {
 
-  test.setTimeout(7 * 60 * 1000); //Extending Test Case timeout to 7 minutes
+  test.setTimeout(10 * 60 * 1000); //Extending Test Case timeout to 10 minutes
 
   await PreProposal(page);   //PreConditions Open URL and Open Mask DAO For Proposal
